+++ crates/vm/backends/revm/mod.rs
@@ -17,6 +17,7 @@ use ethrex_common::types::AccountInfo;
 use ethrex_common::{BigEndianHash, H256, U256};
 use ethrex_storage::{error::StoreError, AccountUpdate};
 
+use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use revm::db::states::bundle_state::BundleRetention;
 use revm::db::AccountStatus;
 use revm::{
@@ -72,8 +73,16 @@ impl REVM {
         let mut receipts = Vec::new();
         let mut cumulative_gas_used = 0;
 
-        for tx in block.body.transactions.iter() {
-            let result = Self::execute_tx(tx, block_header, state, spec_id, tx.sender())?;
+        // Calculate tx senders in parallel
+        let transactions: Vec<_> = block
+            .body
+            .transactions
+            .par_iter()
+            .map(|tx| (tx, tx.sender()))
+            .collect();
+
+        for (tx, sender) in transactions {
+            let result = Self::execute_tx(tx, block_header, state, spec_id, sender)?;
             cumulative_gas_used += result.gas_used();
             let receipt = Receipt::new(
                 tx.tx_type(),